refactor: piece가 타입을 갖도록 변경
