+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,13 +1,14 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
 import chess.domain.Position;
 
 public abstract class Piece {
     protected final Color color;
+    protected final PieceType type;
 
-    protected Piece(Color color) {
+    protected Piece(Color color, PieceType type) {
         this.color = color;
+        this.type = type;
     }
 
     public abstract boolean canMove(Position from, Position to, Piece piece);
@@ -16,5 +17,9 @@ public abstract class Piece {
 
     public Color getColor() {
         return color;
+    }
+
+    public PieceType getType() {
+        return type;
     }
 }
