+++ src/main/java/chess/domain/piece/QueenPiece.java
@@ -1,11 +1,10 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
 import chess.domain.Position;
 
 public class QueenPiece extends Piece {
     public QueenPiece(Color color) {
-        super(color);
+        super(color, PieceType.QUEEN);
     }
 
     @Override
