+++ src/main/java/chess/domain/piece/RookPiece.java
@@ -1,11 +1,10 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
 import chess.domain.Position;
 
 public class RookPiece extends Piece {
     public RookPiece(Color color) {
-        super(color);
+        super(color, PieceType.ROOK);
     }
 
     @Override
