+++ src/test/java/chess/domain/piece/BishopPieceTest.java
@@ -1,6 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
