+++ src/test/java/chess/domain/piece/PawnPieceTest.java
@@ -1,6 +1,5 @@
 package chess.domain.piece;
 
-import chess.domain.Color;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
